seperate SerialTask and DialTask from main.cpp
